Fix a couple more warnings.

+++ src/CEGUIPixmapFont.cpp
@@ -172,7 +172,10 @@ void PixmapFont::load ()
     updateFont ();
 
     char tmp [50];
-    snprintf (tmp, sizeof (tmp), "Succsessfully loaded %d glyphs", d_cp_map.size ());
+    snprintf(tmp, sizeof(tmp),
+             "Succsessfully loaded %d glyphs",
+             static_cast<int>(d_cp_map.size()));
+
     Logger::getSingleton ().logEvent (tmp);
 }
